interfaces with kotlin - very similar to java, just different way of implementing
