+++ app/src/main/java/com/example/helloworld/Inheritance.kt
@@ -9,48 +9,48 @@ package com.example.helloworld
 // Parent class needs to be open to be inherited
 // If I want to make a class non-inheritable use the keyword 'sealed' instead of open
 // Super class, Parent class, Base class -- of ElectricCar class
-open class Car(val name: String, val brand: String) {
-    open var range: Double = 0.0
-
-    fun extendRange(amount: Double) {
-        if (amount > 0) {
-            range += amount
-        }
-    }
-
-    open fun drive(distance: Double) {
-        println("Drove for $distance miles.")
-    }
-}
-
-// Sub class, Child class, Derived class -- of Car Class
-class ElectricCar(name: String, brand: String, batteryLife: Double) : Car(name, brand) {
-
-
-    var chargerType = "Type1"
-    override var range = batteryLife * 3
-
-    override fun drive(distance: Double) {
-        println("Drove for $distance miles on electricity")
-    }
-
-    fun drive() {
-        println("Drove for $range miles on electricity")
-    }
-
-}
-
-fun main() {
-
-    var audiA3 = Car("A3", "Audi")
-    var teslaS = ElectricCar("S-Model", "Tesla", 85.0)
-    teslaS.chargerType = "Type2"
-    teslaS.extendRange(150.0)
-
-    teslaS.drive()
-
-    // Polymorphism
-    audiA3.drive(200.0)
-    teslaS.drive(200.0)
-
-}
+//open class Car(val name: String, val brand: String) {
+//    open var range: Double = 0.0
+//
+//    fun extendRange(amount: Double) {
+//        if (amount > 0) {
+//            range += amount
+//        }
+//    }
+//
+//    open fun drive(distance: Double) {
+//        println("Drove for $distance miles.")
+//    }
+//}
+//
+//// Sub class, Child class, Derived class -- of Car Class
+//class ElectricCar(name: String, brand: String, batteryLife: Double) : Car(name, brand) {
+//
+//
+//    var chargerType = "Type1"
+//    override var range = batteryLife * 3
+//
+//    override fun drive(distance: Double) {
+//        println("Drove for $distance miles on electricity")
+//    }
+//
+//    fun drive() {
+//        println("Drove for $range miles on electricity")
+//    }
+//
+//}
+//
+//fun main() {
+//
+//    var audiA3 = Car("A3", "Audi")
+//    var teslaS = ElectricCar("S-Model", "Tesla", 85.0)
+//    teslaS.chargerType = "Type2"
+//    teslaS.extendRange(150.0)
+//
+//    teslaS.drive()
+//
+//    // Polymorphism
+//    audiA3.drive(200.0)
+//    teslaS.drive(200.0)
+//
+//}
